Update LabelSelect.js
The var instead of this._refreshHandle causes many subscriptions that were never closed and hang the browser for seconds due to memory leaks in case of many reference options en multiple (>5) refreshes of the main object. F5 Solved that in the browser for a while.

The destroy of the this.labelAssignInput is done because the console in 5.9.1 mentions not destroyed comboboxes in my test project.